Moved backup, crypto, opaqueTypes, unsecureRandom and conversion to helpers dir
