+++ src/helpers/backup.ts
@@ -1,5 +1,3 @@
-import { keccak256 } from 'js-sha3';
-import nacl from 'ecma-nacl';
 // @ts-ignore
 import * as base64 from 'base64-arraybuffer';
 // @ts-ignore
@@ -7,9 +5,9 @@ import { generateSecureRandom } from 'react-native-securerandom';
 // @ts-ignore
 import * as utf8 from 'utf8-encoder';
 
-import { hexToByteArray } from './conversion';
-import { Version } from './Version';
-import { encrypt, decrypt } from './cryptoHelpers';
+import { hexToByteArray } from '../helpers/conversion';
+import { Version } from '../Version';
+import { encrypt, decrypt } from '../helpers/crypto';
 
 const header = `-----BEGIN FELFELE BACKUP-----`;
 const headerFields = `
